Delete app to fix error
Once I deployed to Heroku, I got huge error.  I did google and it said different ways of causing it - no corsheaders (which I didnt install at the beginning then I did install when the error but still couldn't fix the problem
Another source said to run the command
152 static files copied to '/Users/srichaii/LambdaSchool/Intro-Django/staticfiles', 474 post-processed. to find what wrong, I did ran and the error said , which I had installed.
Another possibilty I can think of is it might be because I created the app outside the virtual env, so that's why my app doesn;t recognize any of dependencies. (just I guess though).
So I deleted the app through Heroku page and recreate it in pipenv.  Not sure the result but let's see :)