PERSONAL API: FX problem defining endpoints that serve filter data from a specific model.
There is a general model that define all the 'Members'.
The objective is to serve the members foltered by role. Roles = [Directora, Investigadores, etc,...]
As an example if an APP request '/api/lcai/investigadores' the response
is a JSON that contains ONLY the members which role is 'Investigadores'.
I faced several problems trying to get that functionality.
The solution reflected in this commit, even though is not perfect,
achieves the objective.